+++ src/main/java/org/codeiteam3/findex/indexinfo/controller/IndexInfoController.java
@@ -2,10 +2,8 @@ package org.codeiteam3.findex.indexinfo.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.codeiteam3.findex.common.CursorPageResponse;
 import org.codeiteam3.findex.indexinfo.dto.data.CursorPageResponseIndexInfoDto;
 import org.codeiteam3.findex.indexinfo.dto.data.IndexInfoDto;
 import org.codeiteam3.findex.indexinfo.dto.data.IndexInfoSummaryDto;
